pengubahan int ke float

+++ naufal.c
@@ -1,21 +1,21 @@
 #include "naufal.h"
 
-int pertambahan(int nilai1, int nilai2){
-	int temp;
+float pertambahan(float nilai1, float nilai2){
+	float temp;
 	
 	temp = nilai1 + nilai2;
 	return temp;
 }
 
-int pengurangan(int nilai1, int nilai2){
-	int temp;
+float pengurangan(float nilai1, float nilai2){
+	float temp;
 	
 	temp = nilai1 - nilai2;
 	return temp;
 }
 
-int perkalian(int nilai1, int nilai2){
-	int temp;
+float perkalian(float nilai1, float nilai2){
+	float temp;
 	
 	temp = nilai1 * nilai2;
 	return temp;
